requirements.txt
Django==4.2
djangorestframework==3.14.0
mongoengine==0.24.1
python-decouple==3.6
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
